requirements.txt
acbbs==2.0.0
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
celery-progress==0.0.3
certifi==2018.11.29
chardet==3.0.4
colorlog==4.0.2
django-celery-results==1.0.4
etaprogress==1.1.1
idna==2.8
kombu==4.3.0
numpy==1.16.1
pkg-resources==0.0.0
pyModbusTCP==0.1.8
pymongo==3.7.2
pytz==2018.9
requests==2.21.0
scipy==1.2.1
urllib3==1.24.1
vine==1.2.0
